feat: native slider
