+++ src/components/player/controls/ProgressBars/NativeProgressBar.tsx
